Implemented getElucidation in DAL (Testing WIP); Changes to IDialog, create.sql, TestHelper, IElucidation and more

+++ src/main/java/ACQ/IMeeting.java
@@ -79,13 +79,13 @@ public interface IMeeting {
      * Get the number of the meeting
      * @return number of meeting
      */
-    long getNumber();
+    int getNumber();
 
     /**
      * Set the number
      * @param n new number
      */
-    void setNumber(long n);
+    void setNumber(int n);
 
     /**
      * Get the user that created the meeting
